Fixed paging problem
If applied, the paging logic should have an additional logic to check if it exceeded it bounds when changing page sizes. If it does, the system will return the last page of the list instead of returning an empty list.

I coordinated this to HR and they advised that it's better to fix the problem before presenting the code to the stakeholders.